Refactor: Updated members class to match new scheme in app. Changed name from Members to UserDetails.
Got log in working with new scheme
