+++ src/models/user.js
@@ -1,4 +1,3 @@
-import moment from 'moment';
 import md5sum from '../lib/md5sum';
 import parse from '../lib/parse';
 import ParseObject from './parse_object';
@@ -23,27 +22,6 @@ export default class User extends ParseObject {
   // eslint-disable-next-line class-methods-use-this
   get password() { return null; }
   set password(val) { return this.instance.set('password', val); }
-
-  get firstName() { return this.instance.get('firstName'); }
-  set firstName(val) { return this.instance.set('firstName', val); }
-
-  get lastName() { return this.instance.get('lastName'); }
-  set lastName(val) { return this.instance.set('lastName', val); }
-
-  get commitmentAgreementSigned() { return this.instance.get('commitmentAgreementSigned'); }
-  set commitmentAgreementSigned(val) { return this.instance.set('commitmentAgreementSigned', Boolean(val)); }
-
-  get commitmentAgreementSignedDate() { return this.instance.get('commitmentAgreementSignedDate'); }
-  set commitmentAgreementSignedDate(val) { return this.instance.set('commitmentAgreementSignedDate', moment(val).utc().toDate()); }
-
-  get solutioneer101Passed() { return this.instance.get('solutioneer101Passed'); }
-  set solutioneer101Passed(val) { return this.instance.set('solutioneer101Passed', Boolean(val)); }
-
-  get ndaSigned() { return this.instance.get('ndaSigned'); }
-  set ndaSigned(val) { return this.instance.set('ndaSigned', Boolean(val)); }
-
-  get ndaSignedDate() { return this.instance.get('ndaSignedDate'); }
-  set ndaSignedDate(val) { return this.instance.set('ndaSignedDate', moment(val).utc().toDate()); }
 
   static create(wrap = true) {
     const user = new parse.User();
